requirements.txt
flask
python-dotenv
flask-wtf
flask-security
paralleldots